+++ internal/AX25/go.mod
@@ -1,0 +1,3 @@
+module AX25
+
+go 1.23.2
